requirements.txt
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.3.11
python-docx>=0.8.11
pandas>=1.5.0
plotly>=5.15.0
kaleido>=0.2.1
PyPDF2>=3.0.1
pdfplumber>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openpyxl>=3.1.0

# PostgreSQL dependencies
psycopg2-binary>=2.9.9

# Cloudinary for file storage
cloudinary>=1.40.0